make fastpath workers stateful; prep for having more state!
